requirements.txt
torch
transformers
opencv-python-headless
pillow
numpy
roboflow
pandas
evaluate
datasets
scikit-learn
pycocotools
